Move Warning File To Separate Package
